requirements.txt
python-dotenv~=1.0.0
httpx~=0.28.0
rich~=14.0.0
pydantic~=2.11.0
google-api-python-client~=2.169.0
google-auth-oauthlib~=1.2.2
